BackendServices/WebSocketsHandler/requirements.txt
Flask
flask_cors
kafka-python
Flask-SocketIO
eventlet